+++ src/MicroElements.Metadata/Validation/IValidationRule.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) MicroElements. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System;
 using System.Collections.Generic;
 using MicroElements.Functional;
 using MicroElements.Metadata;
@@ -11,7 +10,7 @@ namespace MicroElements.Validation
     /// <summary>
     /// Validation rule.
     /// </summary>
-    public interface IValidationRule
+    public interface IValidationRule : IMetadataProvider
     {
         /// <summary>
         /// Validates property or other aspect of <paramref name="propertyContainer"/>.
@@ -22,14 +21,14 @@ namespace MicroElements.Validation
     }
 
     /// <summary>
-    /// Validation rule with ability to configure validation message.
+    /// Typed validation rule.
     /// </summary>
-    public interface IConfigurableValidationRule : IValidationRule
+    /// <typeparam name="T">Property type.</typeparam>
+    public interface IValidationRule<T> : IValidationRule
     {
         /// <summary>
-        /// Configures validation message.
+        /// Gets property to validate.
         /// </summary>
-        /// <param name="configureMessage">Configuration function.</param>
-        void ConfigureMessage(Func<Message, Message> configureMessage);
+        IProperty<T> Property { get; }
     }
 }